Module tf_serving::examples::mnist

Functions for downloading, reading, and encoding MNIST data.

https://github.com/tensorflow/serving/blob/master/tensorflow_serving/example/mnist_client.py

Structs

DataSet
A vector of (image, label) pair. Retrieve content by indexing: dataset[1].

Functions

extract_images
Extract the images into a vector of ndarray Vec.

extract_labels
Extract the labels into a 1D uint8 numpy array [index].

maybe_download
Download the data from Yann Lecun's website, unless it's already here.

predict_request
Creates a PredictRequest from a given image by setting the model name as "mnist" and encode the image properly.