Struct tf_serving::examples::mnist::DataSet

pub struct DataSet { /* fields omitted */ }

A vector of (image, label) pair. Retrieve content by indexing: dataset[1].

Methods

impl DataSet

fn train<P: AsRef<Path>>(dir: P) -> Result<DataSet>

Loads MNIST training dataset, if it is not present, downloads it.

fn test<P: AsRef<Path>>(dir: P) -> Result<DataSet>

Loads MNIST test dataset, if it is not present, downloads it.

fn len(&self) -> usize

Returns the number of images (or labels).